+++ backend/src/main/kotlin/fr/gouv/dgampa/rapportnav/domain/entities/aem/AEMOutOfMigrationRescue.kt
@@ -5,22 +5,22 @@ import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.action.NavActionEnt
 import fr.gouv.dgampa.rapportnav.domain.utils.AEMUtils
 
 data class AEMOutOfMigrationRescue(
-    val nbrOfHourAtSea: Int? = 0, //1.1.1
-    val nbrOfRescuedOperation: Int? = 0, //1.1.3.
-    val nbrPersonsRescued: Int? = 0, //1.1.4
+    val nbrOfHourAtSea: Double? = 0.0, //1.1.1
+    val nbrOfRescuedOperation: Double? = 0.0, //1.1.3.
+    val nbrPersonsRescued: Double? = 0.0, //1.1.4
 ) {
     constructor(
         navActions: List<NavActionEntity>
     ) : this(
-        nbrOfRescuedOperation = actionRescueEntities(navActions).size,
+        nbrOfRescuedOperation = actionRescueEntities(navActions).size.toDouble(),
         nbrPersonsRescued = getNbrPersonsRescued(actionRescueEntities(navActions)),
-        nbrOfHourAtSea = AEMUtils.getDurationInHours(actionRescueEntities(navActions)).toInt(),
+        nbrOfHourAtSea = AEMUtils.getDurationInHours(actionRescueEntities(navActions)),
     ) {
     }
 
     companion object {
-        fun getNbrPersonsRescued(actionRescues: List<ActionRescueEntity?>): Int {
-            return actionRescues.fold(0) { acc, actionRescue -> acc.plus(actionRescue?.numberPersonsRescued!!) }
+        fun getNbrPersonsRescued(actionRescues: List<ActionRescueEntity?>): Double {
+            return actionRescues.fold(0.0) { acc, actionRescue -> acc.plus(actionRescue?.numberPersonsRescued!!) }
         }
 
         private fun actionRescueEntities(navActions: List<NavActionEntity>): List<ActionRescueEntity?> {
